Move SR parts into `sr` subpackage
