+++ pydicom/__init__.py
@@ -34,7 +34,6 @@ from pydicom.dataset import Dataset, FileDataset
 from pydicom.filereader import dcmread, read_file
 from pydicom.filewriter import dcmwrite, write_file
 from pydicom.sequence import Sequence
-from pydicom.codedict import codes
 
 from ._version import __version__, __version_info__
